making template detail event
