+++ src/components/Atom/Button/style.js
@@ -6,7 +6,7 @@ const buttonStyle = makeStyles((theme) => ({
     fontWeight: 800,
   },
   primary: {
-    background: 'linear-gradient(68.21deg, #EB8402 12.88%, #DA251B 98.79%)',
+    background: 'linear-gradient(68.21deg, #FC6D21 12.88%, #EE3B52 74.92%, #DD2476 98.79%);',
     color: 'white',
   },
   secondary: {
