+++ src/components/index.js
@@ -1,3 +1,4 @@
 export * from './Atom';
 export * from './molecules';
 export * from './organisms';
+export * from './template';
