+++ src/components/template/index.js
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+
+import EventTemplate from './EventTemplate';
+
+export {
+  EventTemplate,
+};
