+++ src/routes/routeList.js
@@ -1,8 +1,8 @@
 import loadable from '@loadable/component';
 
-const ListEvent = loadable(() => import(/* webpackPrefetch: true */ '@page/ListEvent'));
-const DetailEvent = loadable(() => import(/* webpackPrefetch: true */ '@page/DetailEvent'));
-const Login = loadable(() => import(/* webpackPrefetch: true */ '@page/Login'));
+const ListEventPage = loadable(() => import(/* webpackPrefetch: true */ '@/page/ListEventPage'));
+const DetailEventPage = loadable(() => import(/* webpackPrefetch: true */ '@/page/DetailEventPage'));
+const LoginPage = loadable(() => import(/* webpackPrefetch: true */ '@/page/LoginPage'));
 
 const prefix = '/brainspark';
 // const privatePrefix = '/dashboard';
@@ -10,19 +10,19 @@ const prefix = '/brainspark';
 const ROUTES = [
   {
     name: 'Brainspark',
-    Component: ListEvent,
+    Component: ListEventPage,
     path: `${prefix}`,
     isPrivate: false,
   },
   {
     name: 'Brainspark Detail',
-    Component: DetailEvent,
+    Component: DetailEventPage,
     path: `${prefix}/content/:id`,
     isPrivate: false,
   },
   {
     name: 'Login',
-    Component: Login,
+    Component: LoginPage,
     path: `${prefix}/login`,
     isPrivate: false,
   },
